kwaabo/requirements.txt
aiohttp==3.7.3
aiohttp-sse-client==0.1.7
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
crc16==0.1.1
cryptography==3.2.1
Django==2.2.17
django-cors-headers==3.5.0
django-environ==0.4.5
django-model-utils==4.0.0
django-polaris==1.1.5
djangorestframework==3.12.2
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mnemonic==0.19
multidict==5.0.2
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
PyNaCl==1.3.0
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
stellar-base-sseclient==0.0.21
stellar-sdk==2.10.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
whitenoise==5.2.0
wrapt==1.12.1
yarl==1.6.3
